FOO-79 update text rotation if edge is top (#21)
Upon compiling the buttons the border radius of the
is set to the top left and top right of the button.
However, when the button edge is set to the top,
the corners are facing the wrong direction. To fix
this issue the text is rotated whenever the edge is
set tot top.